add user authentication, signup, and post ownership logic

+++ blog/views.py
@@ -1,7 +1,9 @@
 from django.contrib.auth.decorators import login_required
-from django.shortcuts import render, get_object_or_404
+from django.shortcuts import render, get_object_or_404, redirect
 from .models import Post
 from .forms import PostForm
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth import login
 
 # Home page – list all posts
 def home(request):
@@ -16,11 +18,25 @@ def post_detail(request, id):
     post = get_object_or_404(Post, id=id)
     return render(request, 'blog/post_detail.html', {'post': post})
 
+def signup(request):
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save()
+            login(request, user)
+            return redirect('home')
+    
+    else:
+        form = UserCreationForm()
+
+    return render(request, 'registration/signup.html', {'form': form})
+
 @login_required
 def create_post(request):
     if request.method == 'POST':
         form = PostForm(request.POST, request.FILES)
         if form.is_valid():
+            form.instance.author = request.user  # assign author before saving
             form.save()
             return redirect('home')
 
@@ -33,9 +49,13 @@ def create_post(request):
 def edit_post(request, id):
     post = get_object_or_404(Post, id=id)
 
+    if request.user != post.author:
+        return redirect('post_detail', id=id)
+
     if request.method == 'POST':
         form = PostForm(request.POST, request.FILES, instance=post)
         if form.is_valid():
+            form.instance.author = request.user  # assign author before saving
             form.save()
             return redirect('post_detail', id=post.id)
     else:
@@ -46,6 +66,9 @@ def edit_post(request, id):
 @login_required
 def delete_post(request, id):
     post = get_object_or_404(Post, id=id)
+
+    if request.user != post.author:
+        return redirect('post_detail', id=id)
 
     if request.method == 'POST':
         post.delete()
